client/src/main/java/de/njsm/stocks/client/frontend/cli/commands/FaultyCommandHandler.java
/* stocks is client-server program to manage a household's food stock
 * Copyright (C) 2019  The stocks developers
 *
 * This file is part of the stocks program suite.
 *
 * stocks is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * stocks is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */

package de.njsm.stocks.client.frontend.cli.commands;

import de.njsm.stocks.client.exceptions.DatabaseException;
import de.njsm.stocks.client.exceptions.InputException;
import de.njsm.stocks.client.exceptions.NetworkException;
import de.njsm.stocks.client.frontend.cli.Command;
import de.njsm.stocks.client.frontend.cli.service.ScreenWriter;

public abstract class FaultyCommandHandler extends AbstractCommandHandler {

    public FaultyCommandHandler(ScreenWriter writer) {
        super(writer);
    }

    @Override
    public void handle(Command command) {
        handleWithFaultLogger(command);
    }

    protected void handleWithFaultLogger(Command command) {
        try {
            handleInternally(command);
        } catch (DatabaseException e) {
            logDatabaseError(e);
        } catch (InputException e) {
            logInputError(e);
        } catch (NetworkException e) {
            logNetworkError(e);
        }
    }

    protected abstract void handleInternally(Command command) throws
            NetworkException,
            DatabaseException,
            InputException;

}
